requirements.txt
pathlib
json
dotenv
psycopg2-binary